+++ ec2_alb.tf
@@ -4,7 +4,7 @@ resource "aws_lb_target_group" "concourse" {
   protocol = "HTTP"
   vpc_id   = var.vpc_id
 
-  tags = merge({ Name = "${var.prefix}-concourse-backend", role = "loadbalancer" }, var.tags)
+  tags = merge({ Name = "${var.prefix}-concourse-backend" }, var.alb_tags)
 }
 
 resource "aws_lb_target_group_attachment" "concourse" {
@@ -27,7 +27,7 @@ resource "aws_lb" "concourse" {
 
   enable_deletion_protection = false
 
-  tags = merge({ Name = "${var.prefix}-concourse-alb", role = "loadbalancer" }, var.tags)
+  tags = merge({ Name = "${var.prefix}-concourse-alb" }, var.alb_tags)
 }
 
 resource "aws_lb_listener" "concource_http" {