+++ ec2_sgs.tf
@@ -11,7 +11,7 @@ resource "aws_security_group" "GroupLB" {
     cidr_blocks = ["0.0.0.0/0"]
   }
 
-  tags = merge({ Name = "Group-${var.prefix}.coucourse.LB", role = "security" }, var.tags)
+  tags = merge({ Name = "Group-${var.prefix}.coucourse.LB"}, var.sg_tags)
 }
 
 resource "aws_security_group" "GroupWS" {
@@ -26,7 +26,7 @@ resource "aws_security_group" "GroupWS" {
     cidr_blocks = ["0.0.0.0/0"]
   }
 
-  tags = merge({ Name = "Group-${var.prefix}.coucourse.WebServer", role = "security" }, var.tags)
+  tags = merge({ Name = "Group-${var.prefix}.coucourse.WebServer"}, var.sg_tags)
 }
 
 # Rules
@@ -50,7 +50,7 @@ resource "aws_security_group" "RuleGroupLBHttpIn" {
     security_groups = [aws_security_group.GroupLB.id]
   }
 
-  tags = merge({ Name = "Rule-${var.prefix}.coucourse.LB-in-HTTP", role = "security" }, var.tags)
+  tags = merge({ Name = "Rule-${var.prefix}.coucourse.LB-in-HTTP"}, var.sg_tags)
 }
 
 resource "aws_security_group" "RuleGroupWsIn" {
@@ -73,6 +73,6 @@ resource "aws_security_group" "RuleGroupWsIn" {
     security_groups = [aws_security_group.GroupWS.id]
   }
 
-  tags = merge({ Name = "Rule-${var.prefix}.coucourse.WS-in-5432", role = "security" }, var.tags)
+  tags = merge({ Name = "Rule-${var.prefix}.coucourse.WS-in-5432"}, var.sg_tags)
 }